fix(macros): update macro to accept variable arguments
Refactor the `call_dcmi_function` macro to allow for an optional 
variable number of arguments. This change enhances the macro's 
flexibility and usability in different contexts. 

update(structs): modify documentation for SingleDeviceId

Change the documentation comment for `SingleDeviceId` to clarify 
that the ID should not be 65535, improving the accuracy of the 
warning for users.